libconorthedev dependency removed
The classes of libconorthedev have been moved inside the project, because the package is in a different repo and thus, some users may not have Dynastic to install libconorthedev, which makes the package unable to install. Also modified for white text. I might make it support any color, based on the plist, if there is demand for this